Align NBHM behavior with the CHM one. Mutualise tests
